+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # Change Log
 
+## v0.0.3 (2016-05-14)
+**Adicionados**
+* Novo método `check_conds_type()` que verifica os tipos das condições iniciais e de contornos. Se for do tipo function, aplica os valores de `x` ou `y`.
+
 ## v0.0.2 (2016-05-14)
 **Modificados**
 * Laço com a variável `i` em `explicit()` retirado.
 
 ## v0.0.1 (2016-05-14)
 **Adicionados**
-* Classe `Heat1D` com um método numérico explícito `explicit()` para a equação do calor.
+* Classe `Heat1D` com um método numérico explícito `explicit()` para equação do calor.